Chat/requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.0
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
Flask-ReCaptcha==0.5.0

python-socketio==5.5.2
eventlet==0.33.0

WTForms==3.0.1
cryptography==41.0.2
email-validator==1.3.1
gunicorn==20.1.0  # Para despliegue si es necesario
SQLAlchemy==2.0.21  # ORM subyacente para Flask-SQLAlchemy
pyjwt==2.8.0  # Para el manejo de tokens JWT si es necesario
